requirements.txt
customtkinter
Pillow
docx2pdf